moved Page and Pages Component to subdir
