+++ src/components/App.js
@@ -4,8 +4,8 @@ import {BrowserRouter, Redirect, Route, Switch} from 'react-router-dom';
 import Home from './Home';
 import Login from './user/Login';
 import Logout from './user/Logout';
-import Pages from './Pages';
-import Page from './Page';
+import Pages from './page/Pages';
+import Page from './page/Page';
 import {loadPages} from "../actions";
 import {connect} from "react-redux";
 import Footer from './layout/Footer';
